feat: expose assembly-based classification helpers
